requirements.txt
Django==2.0.4
pytz==2018.4
mysqlclient==1.3.12
djangorestframework==3.8.2
